jtag2::recvFrame(): make msglen and its related variables unsigned int so they cannot suddenly become negative.
git-svn-id: svn://svn.code.sf.net/p/avarice/code/trunk/avarice@187 a7e36acc-7495-4a18-91ca-7b9a7881be12

+++ src/jtag2io.cc
@@ -101,8 +101,8 @@ int jtag2::recvFrame(unsigned char *&msg, unsigned short &seqno)
 	sSTART, sSEQNUM1, sSEQNUM2, sSIZE1, sSIZE2, sSIZE3, sSIZE4,
 	sTOKEN, sDATA, sCSUM1, sCSUM2, sDONE
     }  state = sSTART;
-    int msglen = 0, l = 0;
-    int headeridx = 0;
+    unsigned int msglen = 0, l = 0;
+    unsigned int headeridx = 0;
     bool ignorpkt = false;
     int rv;
     unsigned char c, *buf = NULL, header[8];
@@ -233,7 +233,7 @@ int jtag2::recvFrame(unsigned char *&msg, unsigned short &seqno)
     seqno = r_seqno;
     msg = buf;
 
-    return msglen;
+    return (int)msglen;
 }
 
 /*
